Using CMake as the build system instead of .sln

+++ VE/Engine/Window/VWindow.cpp
@@ -1,25 +1,25 @@
-#include "VWindow.h"
-#include <LibsList.h>
-
+#include "VWindow.hpp"
+#include <preprocessing.h>
 //a window that wiill encapsulate the complexity of window managment
-VE::Window::VWindow::VWindow(const wchar_t* Name, unsigned int SizeX, unsigned int SizeY, VAPI Api)
+VE::Window::VWindow::VWindow(const wchar_t* Name, long SizeX, long SizeY, VAPI Api)
 {
 
 	//selection of which api to use the create the window
 	switch (Api)
 	{
-	case VAPI::WINDOWS:
-		this->phwnd = MS_CreateWindow(Name, SizeX, SizeY);
-		update = MS_UpdateWindow;
-		break;
-	case VAPI::OGL:
-		break;
-	case VAPI::VULKAN:
-		break;
-	default:
-		break;
+		case VAPI::WINDOWS:
+			this->phwnd = MS_CreateWindow(Name, SizeX, SizeY);
+			update = MS_UpdateWindow;
+			break;
+		case VAPI::OGL:
+			break;
+		case VAPI::VULKAN:
+			break;
+		default:
+			break;
 	}
-	if (phwnd == nullptr)
+	//check if phwnd in nullptr if so there was a problem creating the window
+	if (phwnd == nullptr) 
 	{
 		printf("Window creation didn't go so well in File: %ls", __FILEW__);
 	}
